Validation

+++ src/main/java/softuniBlog/controller/ArticleController.java
@@ -2,6 +2,7 @@ package softuniBlog.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -44,6 +45,11 @@ public class ArticleController {
         if(!this.articleRepository.exists(id)){
             return  "redirect:/";
         }
+        if(!(SecurityContextHolder.getContext().getAuthentication()instanceof AnonymousAuthenticationToken)){
+            UserDetails principal = (UserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            User entityUser=this.userRepository.findByEmail(principal.getUsername());
+            model.addAttribute("user",entityUser);
+        }
         Article article = this.articleRepository.findOne(id);
         model.addAttribute("article",article);
         model.addAttribute("view","article/details");
@@ -55,6 +61,8 @@ public class ArticleController {
         if(!this.articleRepository.exists(id))
         {return "redirect:/";}
         Article article = this.articleRepository.findOne(id);
+        if(!isUserAuthorOrAdmin(article))
+        {return "redirect:/article/" + id;}
         model.addAttribute("view","article/edit");
         model.addAttribute("article", article);
         return "base-layout";
@@ -66,6 +74,9 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+        if(!isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
         article.setContent(articleBindingModel.getContent());
         article.setTitle(articleBindingModel.getTitle());
         this.articleRepository.saveAndFlush(article);
@@ -79,6 +90,9 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+        if(!isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
         model.addAttribute("article", article);
         model.addAttribute("view","article/delete");
         return "base-layout";
@@ -90,7 +104,15 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+        if(!isUserAuthorOrAdmin(article)){
+            return "redirect:/article/"+id;
+        }
         this.articleRepository.delete(article);
         return "redirect:/";
+    }
+    private boolean isUserAuthorOrAdmin(Article article){
+        UserDetails user = (UserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+        return userEntity.isAdmin() || userEntity.isAuthor(article);
     }
 }
